FQ name unit tests
